noJira: bygningsnummer schema (#123)
* noJira: Bygningsnummer som `z.number().nonnegative()`

* noJira: Oppdatert bygningsnummer i mockdata

+++ src/lib/schema/reports/bygg/byg0011/byggEndring.schema.ts
@@ -15,7 +15,7 @@ import { kontaktpersonSchema } from "../shared/kontaktperson.schema.ts"
 
 export const byggEndringSchema = z
   .object({
-    bygningsnummer: z.number().int().nonnegative().optional(),
+    bygningsnummer: z.number().nonnegative().optional(),
     lopenummer: z.number().int().nonnegative().optional(),
     bygningsendringsKode: kodeSchemaOgTekst(endringsKodeSchema),
     harUfullstendigAreal: jaNeiEnum.optional(),
